refactor: split prompt chat rendering
